# pkg/api/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  APIConfig:
    properties:
      accesslog:
        type: boolean
        x-go-name: AccessLog
      addr:
        type: string
        x-go-name: Addr
      auth:
        $ref: '#/definitions/AuthConfig'
      auther:
        type: string
        x-go-name: Auther
      pathPrefix:
        type: string
        x-go-name: PathPrefix
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  AuthConfig:
    properties:
      password:
        type: string
        x-go-name: Password
      username:
        type: string
        x-go-name: Username
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  AutherConfig:
    properties:
      auths:
        items:
          $ref: '#/definitions/AuthConfig'
        type: array
        x-go-name: Auths
      name:
        type: string
        x-go-name: Name
      type:
        description: inline, file, redis, etc.
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  BypassConfig:
    properties:
      matchers:
        items:
          type: string
        type: array
        x-go-name: Matchers
      name:
        type: string
        x-go-name: Name
      reverse:
        type: boolean
        x-go-name: Reverse
      type:
        description: inline, file, etc.
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  ChainConfig:
    properties:
      hops:
        items:
          $ref: '#/definitions/HopConfig'
        type: array
        x-go-name: Hops
      name:
        type: string
        x-go-name: Name
      selector:
        $ref: '#/definitions/SelectorConfig'
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  Config:
    properties:
      api:
        $ref: '#/definitions/APIConfig'
      authers:
        items:
          $ref: '#/definitions/AutherConfig'
        type: array
        x-go-name: Authers
      bypasses:
        items:
          $ref: '#/definitions/BypassConfig'
        type: array
        x-go-name: Bypasses
      chains:
        items:
          $ref: '#/definitions/ChainConfig'
        type: array
        x-go-name: Chains
      hosts:
        items:
          $ref: '#/definitions/HostsConfig'
        type: array
        x-go-name: Hosts
      log:
        $ref: '#/definitions/LogConfig'
      metrics:
        $ref: '#/definitions/MetricsConfig'
      profiling:
        $ref: '#/definitions/ProfilingConfig'
      resolvers:
        items:
          $ref: '#/definitions/ResolverConfig'
        type: array
        x-go-name: Resolvers
      services:
        items:
          $ref: '#/definitions/ServiceConfig'
        type: array
        x-go-name: Services
      tls:
        $ref: '#/definitions/TLSConfig'
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  ConnectorConfig:
    properties:
      auth:
        $ref: '#/definitions/AuthConfig'
      metadata:
        additionalProperties:
          type: object
        type: object
        x-go-name: Metadata
      tls:
        $ref: '#/definitions/TLSConfig'
      type:
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  DialerConfig:
    properties:
      auth:
        $ref: '#/definitions/AuthConfig'
      metadata:
        additionalProperties:
          type: object
        type: object
        x-go-name: Metadata
      tls:
        $ref: '#/definitions/TLSConfig'
      type:
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  Duration:
    description: |-
      A Duration represents the elapsed time between two instants
      as an int64 nanosecond count. The representation limits the
      largest representable duration to approximately 290 years.
    format: int64
    type: integer
    x-go-package: time
  ForwarderConfig:
    properties:
      selector:
        $ref: '#/definitions/SelectorConfig'
      targets:
        items:
          type: string
        type: array
        x-go-name: Targets
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  HandlerConfig:
    properties:
      auth:
        $ref: '#/definitions/AuthConfig'
      auther:
        type: string
        x-go-name: Auther
      chain:
        type: string
        x-go-name: Chain
      metadata:
        additionalProperties:
          type: object
        type: object
        x-go-name: Metadata
      retries:
        format: int64
        type: integer
        x-go-name: Retries
      tls:
        $ref: '#/definitions/TLSConfig'
      type:
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  HopConfig:
    properties:
      bypass:
        type: string
        x-go-name: Bypass
      hosts:
        type: string
        x-go-name: Hosts
      name:
        type: string
        x-go-name: Name
      nodes:
        items:
          $ref: '#/definitions/NodeConfig'
        type: array
        x-go-name: Nodes
      resolver:
        type: string
        x-go-name: Resolver
      selector:
        $ref: '#/definitions/SelectorConfig'
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  HostMappingConfig:
    properties:
      aliases:
        items:
          type: string
        type: array
        x-go-name: Aliases
      hostname:
        type: string
        x-go-name: Hostname
      ip:
        type: string
        x-go-name: IP
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  HostsConfig:
    properties:
      mappings:
        items:
          $ref: '#/definitions/HostMappingConfig'
        type: array
        x-go-name: Mappings
      name:
        type: string
        x-go-name: Name
      type:
        description: inline, file, etc.
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  ListenerConfig:
    properties:
      auth:
        $ref: '#/definitions/AuthConfig'
      auther:
        type: string
        x-go-name: Auther
      chain:
        type: string
        x-go-name: Chain
      metadata:
        additionalProperties:
          type: object
        type: object
        x-go-name: Metadata
      tls:
        $ref: '#/definitions/TLSConfig'
      type:
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  LogConfig:
    properties:
      format:
        type: string
        x-go-name: Format
      level:
        type: string
        x-go-name: Level
      output:
        type: string
        x-go-name: Output
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  MetricsConfig:
    properties:
      addr:
        type: string
        x-go-name: Addr
      enable:
        type: boolean
        x-go-name: Enable
      path:
        type: string
        x-go-name: Path
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  NameserverConfig:
    properties:
      addr:
        type: string
        x-go-name: Addr
      chain:
        type: string
        x-go-name: Chain
      clientIP:
        type: string
        x-go-name: ClientIP
      hostname:
        type: string
        x-go-name: Hostname
      prefer:
        type: string
        x-go-name: Prefer
      timeout:
        $ref: '#/definitions/Duration'
      ttl:
        $ref: '#/definitions/Duration'
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  NodeConfig:
    properties:
      addr:
        type: string
        x-go-name: Addr
      bypass:
        type: string
        x-go-name: Bypass
      connector:
        $ref: '#/definitions/ConnectorConfig'
      dialer:
        $ref: '#/definitions/DialerConfig'
      hosts:
        type: string
        x-go-name: Hosts
      name:
        type: string
        x-go-name: Name
      resolver:
        type: string
        x-go-name: Resolver
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  ProfilingConfig:
    properties:
      addr:
        type: string
        x-go-name: Addr
      enable:
        type: boolean
        x-go-name: Enable
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  ResolverConfig:
    properties:
      name:
        type: string
        x-go-name: Name
      nameservers:
        items:
          $ref: '#/definitions/NameserverConfig'
        type: array
        x-go-name: Nameservers
      type:
        description: inline, file, etc.
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  Response:
    properties:
      code:
        format: int64
        type: integer
        x-go-name: Code
      msg:
        type: string
        x-go-name: Msg
    type: object
    x-go-package: github.com/go-gost/gost/pkg/api
  SelectorConfig:
    properties:
      failTimeout:
        $ref: '#/definitions/Duration'
      maxFails:
        format: int64
        type: integer
        x-go-name: MaxFails
      strategy:
        type: string
        x-go-name: Strategy
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  ServiceConfig:
    properties:
      addr:
        type: string
        x-go-name: Addr
      bypass:
        type: string
        x-go-name: Bypass
      forwarder:
        $ref: '#/definitions/ForwarderConfig'
      handler:
        $ref: '#/definitions/HandlerConfig'
      hosts:
        type: string
        x-go-name: Hosts
      listener:
        $ref: '#/definitions/ListenerConfig'
      name:
        type: string
        x-go-name: Name
      resolver:
        type: string
        x-go-name: Resolver
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  TLSConfig:
    properties:
      caFile:
        type: string
        x-go-name: CAFile
      certFile:
        type: string
        x-go-name: CertFile
      keyFile:
        type: string
        x-go-name: KeyFile
      secure:
        type: boolean
        x-go-name: Secure
      serverName:
        type: string
        x-go-name: ServerName
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
info:
  title: Documentation of Web API.
  version: 1.0.0
paths:
  /config:
    get:
      operationId: getConfigRequest
      parameters:
      - description: output format, one of yaml|json, default is json.
        in: query
        name: format
        type: string
        x-go-name: Format
      responses:
        "200":
          $ref: '#/responses/getConfigResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Get current config.
      tags:
      - ConfigManagement
    post:
      operationId: saveConfigRequest
      parameters:
      - description: output format, one of yaml|json, default is yaml.
        in: query
        name: format
        type: string
        x-go-name: Format
      responses:
        "200":
          $ref: '#/responses/saveConfigResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Save current config to file (gost.yaml or gost.json).
      tags:
      - ConfigManagement
  /config/authers:
    post:
      operationId: createAutherRequest
      parameters:
      - in: body
        name: data
        schema:
          $ref: '#/definitions/AutherConfig'
        x-go-name: Data
      responses:
        "200":
          $ref: '#/responses/createAutherResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Create a new auther, the name of the auther must be unique in auther
        list.
      tags:
      - ConfigManagement
  /config/authers/{auther}:
    delete:
      operationId: deleteAutherRequest
      parameters:
      - in: path
        name: auther
        required: true
        type: string
        x-go-name: Auther
      responses:
        "200":
          $ref: '#/responses/deleteAutherResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Delete auther by name.
      tags:
      - ConfigManagement
    put:
      operationId: updateAutherRequest
      parameters:
      - in: path
        name: auther
        required: true
        type: string
        x-go-name: Auther
      - in: body
        name: data
        schema:
          $ref: '#/definitions/AutherConfig'
        x-go-name: Data
      responses:
        "200":
          $ref: '#/responses/updateAutherResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Update auther by name, the auther must already exist.
      tags:
      - ConfigManagement
  /config/bypasses:
    post:
      operationId: createBypassRequest
      parameters:
      - in: body
        name: data
        schema:
          $ref: '#/definitions/BypassConfig'
        x-go-name: Data
      responses:
        "200":
          $ref: '#/responses/createBypassResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Create a new bypass, the name of bypass must be unique in bypass list.
      tags:
      - ConfigManagement
  /config/bypasses/{bypass}:
    delete:
      operationId: deleteBypassRequest
      parameters:
      - in: path
        name: bypass
        required: true
        type: string
        x-go-name: Bypass
      responses:
        "200":
          $ref: '#/responses/deleteBypassResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Delete bypass by name.
      tags:
      - ConfigManagement
    put:
      operationId: updateBypassRequest
      parameters:
      - in: path
        name: bypass
        required: true
        type: string
        x-go-name: Bypass
      - in: body
        name: data
        schema:
          $ref: '#/definitions/BypassConfig'
        x-go-name: Data
      responses:
        "200":
          $ref: '#/responses/updateBypassResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Update bypass by name, the bypass must already exist.
      tags:
      - ConfigManagement
  /config/chains:
    post:
      operationId: createChainRequest
      parameters:
      - in: body
        name: data
        schema:
          $ref: '#/definitions/ChainConfig'
        x-go-name: Data
      responses:
        "200":
          $ref: '#/responses/createChainResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Create a new chain, the name of chain must be unique in chain list.
      tags:
      - ConfigManagement
  /config/chains/{chain}:
    delete:
      operationId: deleteChainRequest
      parameters:
      - in: path
        name: chain
        required: true
        type: string
        x-go-name: Chain
      responses:
        "200":
          $ref: '#/responses/deleteChainResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Delete chain by name.
      tags:
      - ConfigManagement
    put:
      operationId: updateChainRequest
      parameters:
      - in: path
        name: chain
        required: true
        type: string
        x-go-name: Chain
      - in: body
        name: data
        schema:
          $ref: '#/definitions/ChainConfig'
        x-go-name: Data
      responses:
        "200":
          $ref: '#/responses/updateChainResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Update chain by name, the chain must already exist.
      tags:
      - ConfigManagement
  /config/hosts:
    post:
      operationId: createHostsRequest
      parameters:
      - in: body
        name: data
        schema:
          $ref: '#/definitions/HostsConfig'
        x-go-name: Data
      responses:
        "200":
          $ref: '#/responses/createHostsResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Create a new hosts, the name of the hosts must be unique in hosts list.
      tags:
      - ConfigManagement
  /config/hosts/{hosts}:
    delete:
      operationId: deleteHostsRequest
      parameters:
      - in: path
        name: hosts
        required: true
        type: string
        x-go-name: Hosts
      responses:
        "200":
          $ref: '#/responses/deleteHostsResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Delete hosts by name.
      tags:
      - ConfigManagement
    put:
      operationId: updateHostsRequest
      parameters:
      - in: path
        name: hosts
        required: true
        type: string
        x-go-name: Hosts
      - in: body
        name: data
        schema:
          $ref: '#/definitions/HostsConfig'
        x-go-name: Data
      responses:
        "200":
          $ref: '#/responses/updateHostsResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Update hosts by name, the hosts must already exist.
      tags:
      - ConfigManagement
...
produces:
- application/json
responses:
  createAutherResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  createBypassResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  createChainResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  createHostsResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
...
  deleteAutherResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  deleteBypassResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  deleteChainResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  deleteHostsResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
...
  getConfigResponse:
    description: successful operation.
    headers:
      Config: {}
    schema:
      $ref: '#/definitions/Config'
  saveConfigResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  updateAutherResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  updateBypassResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  updateChainResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  updateHostsResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
...
schemes:
- https
- http
securityDefinitions:
  basicAuth:
    type: basic
swagger: "2.0"